+++ lib/aasm/errors.rb
@@ -1,4 +1,5 @@
 module AASM
   class InvalidTransition < RuntimeError; end
   class UndefinedState < RuntimeError; end
+  class NoDirectAssignmentError < RuntimeError; end
 end