+++ spec/models/persistence.rb
@@ -48,7 +48,22 @@ end
 class NoScope < ActiveRecord::Base
   include AASM
   aasm :create_scopes => false do
-    state :ignored_scope
+    state :pending, :initial => true
+    state :running
+    event :run do
+      transitions :from => :pending, :to => :running
+    end
+  end
+end
+
+class NoDirectAssignment < ActiveRecord::Base
+  include AASM
+  aasm :no_direct_assignment => true do
+    state :pending, :initial => true
+    state :running
+    event :run do
+      transitions :from => :pending, :to => :running
+    end
   end
 end